+++ src/main/java/io/katamari/handler/HelloWorld.java
@@ -1,6 +1,5 @@
 package io.katamari.handler;
 
-import io.netty.channel.ChannelInboundMessageHandlerAdapter;
 import io.netty.channel.ChannelHandlerContext;
 
 import io.katamari.Env;
@@ -10,5 +9,10 @@ public class HelloWorld extends InboundMessageHandler {
   @Override
   public void messageReceived(ChannelHandlerContext ctx, Env env) throws Exception {
     env.getResponse().end("Hello World");
+  }
+
+  @Override
+  public void endMessageReceived(ChannelHandlerContext ctx) throws Exception {
+    ctx.flush();
   }
 }