+++ src/main/java/io/katamari/handler/InboundMessageHandler.java
@@ -8,12 +8,7 @@ import io.katamari.Env;
 public class InboundMessageHandler extends ChannelInboundMessageHandlerAdapter<Env> {
   @Override
   public void messageReceived(ChannelHandlerContext ctx, Env env) throws Exception {
-    ctx.nextInboundMessageBuffer();
-  }
-
-  @Override
-  public void endMessageReceived(ChannelHandlerContext ctx) throws Exception {
-    ctx.flush();
+    ctx.fireInboundBufferUpdated();
   }
 
   @Override