+++ src/test/java/io/katamari/handler/UriDecoderTest.java
@@ -17,6 +17,7 @@ import static io.netty.handler.codec.http.HttpMethod.*;
 
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.handler.codec.http.DefaultFullHttpRequest;
+import io.netty.buffer.MessageBuf;
 
 import io.katamari.Env;
 import io.katamari.env.Request;
@@ -25,10 +26,19 @@ import io.katamari.handler.UriDecoder;
 @RunWith(MockitoJUnitRunner.class)
 public class UriDecoderTest {
   @Mock private ChannelHandlerContext context;
+  @Mock private MessageBuf<Object> buf;
   
   private UriDecoder handler = new UriDecoder();
   private DefaultFullHttpRequest request = new DefaultFullHttpRequest(HTTP_1_1, GET, "/path?id=1");
   private Env env = new Env(context, request);
+
+  @Before
+  public void initialize() {
+    MockitoAnnotations.initMocks(this);
+    when(context.nextInboundMessageBuffer()).thenReturn(buf);
+    when(buf.unfoldAndAdd(anyObject())).thenReturn(true);
+    when(context.fireInboundBufferUpdated()).thenReturn(context);
+  }
 
   @Test
   public void exposesPathInRequest() throws Exception {